Factor out query style normalization
